Update ComponentsCore (add load help-doc for plugins; small refactoring)
